Give matched file relative to base.
Instead of returning an absolute path to the matched file, return a path
relative to the base directory from which the match was performed.

This match is performed using `path.relative`. Results are not expected
to be identical to results using bash for `fnmatch`. A match in a
directory that is not a subdirectory of the base is going to be whatever
it's going to be. We'll tell you how to get there.

Closes #11.