First pass at argument parsing, usage.
Took a first pass at argument and usage parsing. Converted a getopt
function I'd written, rather than writing something less cryptic.

Not sure how to output usage, so I raise an exception. The user can
figure out how to get the usage message to the console before process
exit. I've been having a time of it on Windows. It seems like the sort
of thing that will change a lot before 1.0, so I'll leave it to the
application developer to wrestle with buffers and drainage.

Closes #4.