Remove `sort` from flatten.
No longer sorting keys nor values in `flatten`. Able to test in Node.js
because V8 preserves insertion order. Not supposed to rely on this, but
it is the way it is, and we're testing. If the test breaks, we'll
rethink using sort.

Don't you go relying on the sort order in your applications.

Closes #17.