Add `$given`, a list of given parameters.
Created a `$given` array. An array of parameters given, as opposed to
merely present in the map. This way the application developer does not
have to filter out keys that start with dollar signs, nor test to see if
an array is or is not empty to see if was a parameter actually
specified.

Closes #16.