Internationalization.
Multiple languages supported simply by writing multiple usage messages.

Write a usage message in the appropriate language. Select a usage
message by passing in the locale code. The locale code will match based
on the word used to start the description of the command line.

Tempted to renamed the library `usage`.

Also, flattening arugments passed to `parse`, so it's easier to prepend
or append extra arguments.

Closes #18.
Closes #8.