Wrap program in try/catch block.
Extended the parse method to accept a main program function that will be
run wrapped in a try/catch block. The main program function can throw an
`Error`. If the error matches a validation message, the error is
displayed along with the usage message.

This commit includes a few broken tests. I'm abandoning the notion of
determining the locale of the usage message by detecting the language of
the usage message. The next commit will introduce a structural markup
language; headers and footers for the usage messages. The tests that
fail touch code that is about to be rewritten.