Provide format to `abended` function.
Instead of providing the message string and the components of the
message format as arguments, add them to an exception and pass the
exception to the `abended` function. That way, the `abended` function
can simply throw or forward the exception.

With this commit, I've placed the format components into a `format`
object to keep from adding too much to the `Error` object namespace.

Closes #35.