Correctly set exit code.
In 0.10.40 there was no way to passively set the exit code. If you
wanted to set the exit code it could only be set by calling the `exit()`
function, which would exit immediately. In Arguable, I attempted to do
this using an `"exit"` handler, but doing so defeats the default action
of uncaught exceptions which is to print a stack trace.

Therefore, I'm no longer going to support 0.10. 0.12 let's you set the
exit code passively for an orderly exit. We'll do that instead. If the
user does not want to exit with the default error for an uncaught
exception, then they should catch it and exit in an orderly fashion.