Sketch of destructible based Arguable.
Convert the parameters test to use a Destructible based Arguable. This
test should really be called an options test because we're now calling
these things options in our refactored thinking.

The idea is that there are services you might want to mock that you
ordinarily import with something like `require('lib')`. You can
configure an option with a `lib: require('lib')` but in testing pass in
a mocked lib. This keeps you from having branches in your main module
that can only be reached during testing.

Keep in mind that testing the main module is difficult because it is
supposed to just run when you source it. Arguable is designed to give
you a function that it will run if sourced as the main module or export
otherwise so you can get your hands on it and unit test it.