Implement `Child.exit` and `Child.destroy`.
When invoking a main module in testing we're going to want to signal a
program shutdown but I no longer want to simulate `SIGTERM` through the
signal handling. That was useful when the main module was going to
define its own shutdown and cleanup but now it is going to rely on
Destructible to shutdown and cleanup. We don't need to test Arguable's
signal hooks when we're testing our dear user's main module.

We know we're going to call `Destructible.destroy` so we expose a
`Child` object that does this and that returns the return value of
`Destructible.completed.wait` through `Child.exit`.