Sketch of new signal debugging.
For starters we have an option named `$signals` that is an
`EventEmitter` to use in lieu of `process` so you can send signals from
your unit test.

There are three default signal handlers in Node.js that do not do what
we want. `SIGINT`, `SIGTERM` and `SIGHUP` all exit Node.js immediately
abandoning any work left in the work queue. If you assign a signal
handler yourself then this default behavior is defeated.

We want to unwind our work queue using Destructible so by default we
hook `SIGINT` and `SIGTERM` to `Destructible.destroy`. By default we
assign a no-op handler to `SIGHUP` because I've always thought of it to
be a signal that tells a program to reload its configuration, not one
that tells the program to shutdown. If I'm not actually going to reload
your configuration then I want `SIGHUP` to do nothing.

You can override this behavior with a `$trap` option. The value of
`$trap` is either `true` for default traps, `false` for no traps, or an
object that maps signals to our behaviors of `"destroy"`, `"swallow"` or
`"default"`. If you want to set an actual signal handler you can do that
in your program.

Need to expose the actual signal `EventEmitter` in Arguable.