Capture errors on initialization and runtime.
Looks like I removed the logic in Destructible that would report an
initialization error. It does when you do not provide a callback and
instead have Destructible create a callback, but it does not if you
provide a callback.

There where now two places where errors where reported, at
initialization and at exit, so I thought about having a `ready` latch
that would return the return value from the initialization function
instead of or in addition to providing a callback.

Instead I fudged what Destructible does. If initialization errors,
instead of returning the error I feed it to an ephemeral
`initialization` monitor callback. If there is no error I just callback
the monitor so it exits. The monitor will not inject anything into the
return values because it is ephemeral.

This is probably okay. The root has special behavior. If I find myself
doing this more than once I rethink what is going on in Destructible.

I believe I wanted to leave it up to the caller to decide how this error
should rejoin the Destructible report which is essentially a tree. Here
I'm putting into Destructible explicitly. Probably okay.