Expose `Arguable.identifier`.
Expose the identifier used when creating the `Destructible`. This is
primarily to test it, though.

Seems like we might want to use an identifier based on argument values
sometimes, so maybe there is a way to set the identifier in the main
function. For specialized child processes identifier arguments could
probably be set from environment variables. For other applications you
might want to use a shortened program name.

Maybe just the program name should be the default and not the full path.