Implement new option merging.
We're not going to have a tree of options. We're going to mix our
special options with additional user options. Our special options will
be prefixed with a `$` dollar sign. We'll still share them with the
user, not puck them, but it will make it easier to see at a glance what
is going to have an effect on Arguable and what are user override-able
options.