Implement `$isMainModule` and `$untrap`.
When `$untrap` is `true` Arguable will remove the signal handlers it set
when the program exits. If not explicitly set it is set to `false` when
the user module run as the main module, `true` when the user module is
called directly, probably during testing, so that we don't pollute
`process` with extraneous signal handlers during testing.

Also implemented `$isMainModule` to set whether or not
`Arguable.isMainModule` is `true` or `false`. At this juncture, this is
primarily to test the above `$untrap` setting behavior, but at some
point in the history of Arguable it was necessary or at least I felt it
was necessary so it's not a complete hack.

I was able to delete a pseudo-`process` accessor -- the
`Arguable.mainModule` accessor. So the pruning is underway.

Restored test coverage to 100% so I can now start pruning in earnest.