Remove `exitCode` getter.
`Arguable` is no longer a mock process. We're going to get an exit code
from a wicked exception instead.

Wait, that means that the response that the return values from the
`exit` function of the child are not going to happen. Destructible will
lose track of its return values if we raise an exception.

Suppose then that could be an `arguable.exit(0)` that will record the
exit code and then call `Destructible.destroy`. Either that or we have
no plan for the return values from `Destructible`.