Sketch of abnormal exit.
Still want to the ability to crash the program and stop going forward if
there are exceptions, but if those exceptions are expected as part of
the program, to report them to standard error without a stack trace and
other spew. This preserves validation and even means we can validate if
outside of the initial main function. Currently that's not possible, or
maybe it is just hard to get right.

At the same time, I might want to continue to use the exit return values
coming from Destructible. Leaning toward using `Arguable.exit(code)` to
mean unwind Destructible and return the given code while
`Arguable.raise(code)` will raise an exception and return the code. It
can include a message formatting instruction.

We keep `help` and `helpIf` as well. They've always worked since I
always call them immediately.

Now I'll be able to load configuration files and validate them, or check
that the network is reachable, without spewing stack traces.