Set standard I/0.
We're going to make our dear user set dummy standard I/O streams instead
of creating them on our dear user's behalf. The `$stdout`, `$stderr` and
`$stderr` arguments are visible in the options returned after
construction. This might be annoying if we want to assert that raising
an exception with `Arguable.raise` did correctly format the messages
written to standard error, but internationalization probably shouldn't
be unit tested. Programmatic output should be.

It is not to terribly painful to make the dummy standard I/O streams
visible to both the try and catch blocks of a Cadence exception if you
really want to verify the contents of standard error on an exception.