Create a `Messenger` utility.
It will be the case that I'll want to unit test Node.js IPC. When I do,
I'm going to create this utility in the unit test each time, so I may as
well create it here.

But unlike the Arguable of yore, the new Arguable will not infer the
need for a `send` function and provide an interface. This is an extra
option that our dear user will set and it will not be special in any
way. It will not be specified with a special `$`-prefixed option. Our
user will have to chose to use this object in testing and use `process`
for the same option in the definition for use when run as the main
module.