Simplify delegate.
Going to keep delegate since I do use it a lot. It needs to pass the
arguables options and configured I/O to the delegate. Some mixture of
merged defintion and invocation options plus defaults. Maybe the
generated options hash should contain all the options necessary to
correctly delegate.

At this point all of the test programs are Destructible based so I can
work on removing the old invocation logic. Test coverage is down because
of the dead non-Destructible branch.