Make `options` a property of `Arguable`, `Child`.
Instead of passing in `options` as an extra argument, make them a
property of both the `Arguable` and the `Child`. For a direct invocation
the `options` property is the same for the `Arguable` inside the main
function and the `Child` returned by the invocation used to destroy the
`Destructible` and obtain the exit code.