Delegation requires module argument.
Delegation is such a special thing. So far, it has always dynamically
loaded the child module based on the first argument not parsed out of
`argv` by `getopt`. Not going to offer a lot of options until I need
them.

This is always a very simple module too. They are very specific. They
all look the same. They are relatively easy to test. I don't feel like I
need to make the module available to Arguable to implement this.