Test Destructible as default.
All of our fixtures specified an options argument to indicate that they
where Destructible based so that we where not testing the case when no
options are specified. Nice Destructible is now the default we've
removed options arguments when their only property is the
`$destructible` flag specifying the default value of `true`.