Implement pseudo-pipes.
I want the ability to mock pipes in Arguable main methods. It is
something I've had to repeatedly test by setting up a TCP server in a
test and connecting to the socket created by the TCP server.

In our options `$pipes` is an object keyed by file descriptor. The
values are either arguments to a `net.Socket` constructor for production
or a class derived from `stream.Stream` for mocking. The mock stream
will probably be a `stream.PassThrough` in testing.