No `Destructible.destroy.bind` with signals.
Cannot use `Destructible.destroy.bind` with signals because signal
handlers send arguments and `Destructible.destroy` asserts that it
receives no arguments. Perhaps it shouldn't as this sort of binding
seems sensible, but `Destructible.destroy` used to optionally receive an
`error` and when I removed it I added the assertion to ensure nothing
was depending on that behavior.