`Arguable.require` only does module load.
I don't want to have to repeat the load test in every module that is
designed to delegate, but I want to punt decisions on how to propagate
the configuration to the child. It seems like I'd have to write a
bunch of logic to create an options array that accounts for traps
already being set, whether standard I/O was default or specified, and
then how does the delegator override this logic if it wants to?

Why implement this at all? As noted, I don't want to have test and reset
a failed delegate load. Although, that is pretty simple, you just need
to invoke with a command that is missing.

This feature is now dubious.