Destructible waits on itself to construct.
If the caller creating a child destructible to do something like invoke
a sub command they may pass in that destructible with the constructor
for that destructible's enclosed monitor. In this case the provide
destructible is waiting on itself. From the outside there is no good way
to fix this, to create a destructible and not have to wait on a return
value, or at least it looks really confusing to me and what I created
here works.

One destructible named `run` encloses two destructibles named
`initialize` and `main`. The nesting is a bit much, though. I'd like to
come back and see if there is another way.