Remove double nested invocation of main method.
For applications that are not performing some form of delegation having
the nested `main` and then `run` is going to look silly. `main` is
already probably too much but we do need it to merge our error.

I did remove the ability to specify a destructible. Calling a
subordinate module and having it create it's own destructible works just
fine with `abend` propagation because the error is stripped to abend
before it is passed to `child.exit.unlatch`. Thus, the caller gets just
the `abend` exception and does not have to extract it from a
Destructible tree of exceptions.

Occurs to me that a no-destructible constructor could pass the
destructible and arguable as arguments just straight. Or maybe delegates
are just functions that expect a destructible and an arguable? No
because delegates are supposed to parse arguments.

In any case, if you could just call a delegate directly, or maybe if you
do implement `$destructible`, we just call the delegate directly. We
assume that someone else is going to merge the initialization error.