Configure scram from options and program arguments.
Added the ability to configure the Destructible scram from both the
options object and from the program arguments. There is now an optional
`$scram` option. If it is a number it is used as the scram timeout for
the `Destructible.` If it is a string, the integer value of the program
argument named in the string is used as the scram timeout, otherwise the
default timeout if the argument is not given. If it is an object with
a single property with an integer, the integer value of the program
argument with the same name as the property is used as the scram
timeout, otherwise the value for the property is used.

The program argument is validated as a non-zero integer using
`Arguable.assert` and a user-friendly exception is raised if validation
fails.

Also, changed the default timeout to ten seconds instead of
`Destructible`'s one second.