Add an `exited` signal to `Arguable`.
This has become necessary to exit a cluster child. The call to
disconnect from the parent must come last because the process will exit
without waiting for a `Destructible` to complete once all the servers
and their sockets have closed. Yet another variation on shutdown
behavior. Olio will use this signal to disconnect on the next event loop
after `Destructible` has completed.

Which means that we've designed this to be synchronous and to ensure
that `process.exitCode` is set correctly. If there is an error, any
additional work will not occur because we are going to throw that error
as an uncatchable exception.